+++ NEWS.md
@@ -19,6 +19,8 @@
 	* `set_style_options()`: Sets the options from `excel_output_style()` and `number_format_style()` to a global environment so that every function which is capable of outputting styled Excel workbooks can use them without passing the `style` parameter every time individually. (06.01.2026)
 	* `get_style_options()`: Print the currently set global options. (06.01.2026)
 	* `reset_style_options()`: Set global style options to default values of `excel_output_style()` and `number_format_style()`. (06.01.2026)
+	* `close_file()`: A simple, more readable wrapper for setting file parameter to NULL. (07.01.2026)
+	* `set_variable_labels()`, `get_variable_labels()`, `set_statistic_labels()`, `get_statistic_labels()`, `set_print()`, `get_print()`, `set_monitor()`, `get_monitor()`, `set_na.rm()`, `get_na.rm()`: Additional global setters and getters. (07.01.2026)
 
 ### New functionality
 * `inverse()`: Now supports variable names written without quotation marks. (21.12.2025)
@@ -28,6 +30,8 @@
 * `args_to_char()`: Now converts the contents of a given argument to a character vector. (02.01.2026)
 * `excel_output_style()`: Parameter `file` is now split up into `save_path` and `file` (meaning just the file name + extension). (06.01.2026)
 * `any_table()`, `frequencies()`, `crosstabs()`, `export_with_style()`: The `style` parameter is now set to the new global styling options as default instead of `excel_output_style()`. (06.01.2026)
+* `any_table()`, `frequencies()`, `crosstabs()`: The `var_labels` and `stat_labels` parameter is now set to the new global options as default. (07.01.2026)
+* In general: New global options for `print`, `monitor` and `na.rm` have been implemented into the functions capable of using them. (07.01.2026)
 
 ### Removed
 * `split_by_var()` and `split_by_condition()`: See comment under 'New functions'. (30.12.2025)
@@ -42,6 +46,7 @@
 * `frequencies()`: 'formats' parameter was missing a '=' in examples. (02.01.2026)
 * `summarise_plus()`: Group percentages with nesting = "all" or "single" and na.rm = TRUE are now computed as intended. (02.01.2026)
 * `handle_cell_styles()`: Set apply_font and font_id in a save way to prevent warnings. (04.01.2026, thanks to @JanMarvin)
+* `summarise_plus()`: Removed conversion to numeric values before applying formats, which could lead to not matching formats, if a numeric value was intentionally stored as character value. (07.01.2026)
 
 ### Optimization
 * `summarise_plus()`: Now uses faster `collapse::na_omit` for NA removal. (27.12.2025)